requirements.txt
fastapi
uvicorn
websockets
pydantic
openai
python-dotenv
langchain
langchain_openai
langchain-community
langgraph
qianfan
tencentcloud-sdk-python
websocket-client
httpx_sse
python-multipart
cos-python-sdk-v5
PyJWT
sqlalchemy
redis
langfuse
langsmith
llama-index
pymupdf ==1.24.7
llama-index-readers-json
llama-index-embeddings-huggingface
llama-index-embeddings-dashscope
llama-index-embeddings-jinaai
FlagEmbedding
peft
pymilvus
llama-index-vector-stores-milvus
llama-index-vector-stores-chroma
llama-index-postprocessor-jinaai-rerank
llama-index-retrievers-bm25
llama-index-llms-groq
torch==2.3.1
torchvision==0.18.1
timm
shapely==2.0.1
llama-index-embeddings-ollama
toml
llama-index-callbacks-langfuse
zhipuai




